Physics

+++ src/hs/sceneobject/Circle.java
@@ -20,7 +20,7 @@ public class Circle extends Polygon
         super(gravity, origin, xScale, yScale);
     }
 
-    public Circle(Vector2 gravity, Point2D.Float origin, int xScale, int yScale, boolean hasPhysics, int numSides)
+    public Circle(Vector2 gravity, Point2D.Float origin, float xScale, float yScale, boolean hasPhysics, int numSides)
     {
         super(gravity, origin, xScale, yScale, hasPhysics, numSides);
     }
